add aide repo, add data dir in infra, add aides velo controller, import json instead of readd yaml file
