+++ package.json
@@ -33,7 +33,6 @@
     "@nestjs/platform-express": "^9.0.0",
     "@nestjs/swagger": "^6.3.0",
     "@prisma/client": "^4.15.0",
-    "js-yaml": "^4.1.0",
     "publicodes": "^1.0.0-beta.70",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.2.0",
